For position, use create time and fractional index (#36)
This solves a problem we had in thoughtstream, where when you create a note at the top didn't seem to stay at the top. If you create a bunch of notes at the top on an online client, then create one note at the top on an offline client (which doesn't have those new notes yet), when you go online, that note you created while offline will be positioned below the notes you created earlier while online. This is unintuitive behaviour. When a user creates something at the top, they expect it to be roughly at the top after sync.

The way to do this is by having a part of integer part of the index. When you create a note at the top, you use the current time as the integer part. When sorting, you sort by this int first, which will put more recently created notes above. When you want to be a note underneath another, you use the note above's integer part, then generate a fractional index that places it below the note above.

Now when we sort by the integer part and then the fractional index, we get the ordering we expect.