feat: at-sign context menu new styling
Just some spacing fixes and color adjustments.

I also added the home icon at the beginning of the path, but I'm not sure if that was the intended look. I heard Jacob mention it when he pointed to the "Create New Node" button.

We can iterate based on this, right?