fix: solve icon shifting
So far, this is my best attempt at fixing the kind of 'jumping' Jacob pointed out. I moved the chevron to the far left because it's the only icon that isn't persistent in the UI. I preferred it closer to the node, though. I'll see if I can come up with a better solution later.