fix: Make x button in search bar clear it
The x button was missing a setInput("") so wasn't clearing it.

But we can just get rid of the extra input state, and use the search query
instead. There was a comment about the component not updating in response
but that was just because the search bar was missing an observer wrapper
function.