Remove editor dependency from graph (#221)
For another PR I was trying to write unit tests and I hit due to importing some css files. There's supposed to be a config in our jest to accomodate css files, but I couldn't figure out why they weren't working. But anyways... the file I was testing shouldn't have included css anyways.

It turns out the graph is dependent on some editor helpers which are dependent on css files. It makes sense that the editor needs to know about ui and css. There shouldn't be a dependency between the graph and the ui though.

In this PR, I refactor the split node method to get rid of that dependency. This method still needs to be cleaned up a lot, but this accomplishes what I want for now.