Build tree at top of view (ENT-3592) (#214)
The main change is moving the responsibility of building the visible tree inside the Tree class. Previously, only the root object was passed into the react components, and it was the responsibility of those components to recursively grab related objects, filter, etc. Now the entire nested tree is constructed in one go at the top, and it flows down through rendering.

The concept of a TreeNode is introduced. Graph objects don't have a single parent within the graph, but when it's rendered inside an outline, it does. The tree node wraps the object, contains a pointer to it's parent, and a list of "children" tree nodes.

I think I've got most basic functionality working again e.g. pinned, filters, expansions. I'll do a follow on PR to do some more clean up and fix any missing functionality.