Refactor split (#232)
Moved responsibility of node splitting into the Tree class. Also generally tried to make it simpler and more readable. I removed the bundle handling for now and will maybe re-introduce in ENT-3673.

Also added an argument to FractionalPositionedList, GraphStore.pinRelation, and GraphStore.addChildNode which allows caller to specify a position. This simplifies code in a few places where we'd need to first create/add something and then move it.

part of ENT-3671