Foundational pieces of sync system (ENT-3678) (#251)
The basic idea here:
* Every transaction generates a `SyncTask`, which includes data to sync
and a callback to undo the transaction if sync fails
* We process all queued `SyncTask`s every 500ms. If sync fails on any
given task, then we undo everything that came after the task that failed
to sync (moving backwards) to revert us to the state right before the
failure
* The backend server streams out all successful transactions to all
clients via Websocket. Clients keep track of the ones they've sent and
ignore them, only processing updates from other clients
* We have per-object version numbers, which we increment each time a
`GraphStore` transaction updates those objects. When the backend goes to
apply a transaction, it checks that the updated object(s) are
incrementing `version` by exactly 1, and rejects otherwise, which will
be a cue to the client to pull new data.

Big things still TODO that will be addressed in follow ups:
- [ ] Do version checks within SQL `select` for upserts
- [ ] Re-pull logic when clients transactions get rejected
- [ ] Some kind of batching of node update transactions to make API less
chatty
- [ ] A whole lot of bug fixes
- [ ] Unit tests