Add back the "blue underline" with  a new UI (#271)
Summary
- Previously, we had a view of nodes we called "blue underline" and a
method "treat as link" for determining the view. In this PR, we're
calling the view local vs reference, and have isLocal computed property
to determine which it is. If a node is "local" we render editable
content, otherwise we render it as a pill looking reference.
- Add component to TreeNodeReference render nodes as references
- Clean up mentions plugin
- Convert GraphObject interface into an abstract class which implements
many methods that are shared across graph nodes and relations.

Completes ENT-3652