fix bug where you'd lose focus after assigning a relation type using colon in the editor (#278)
When we assign a relation type on colon, we remove the preceding text
from the current node. This update to the graph then gets reflected in
the editor by SyncWithGraphPlugin. When we update the editor content,
the focus gets lost. The BindFocusToTreePlugin would try to refocus, but
that happened _before_ the editor's content got updated. We address this
by refocusing the editor within the SyncWithGraphPlugin, after we update
the content.