Misc routing changes (#289)
## Summary

The main change is fixing an issue where you'd land on the page, click
on a bullet, and the whole page would appear to reload. This was cause
we were using different routes for `/app/page` and `/app/[[...path]]`,
and the later wasn't prefetched. Now we just redirect right at the
start, which avoid a janky experience after loading.

Other changes:
- Remove split view. This was needed during prototyping but isn't
actually necessary from a product perspective.
- Rename outline view type to graph. This is in preparation for the
outline (nested bullets) being one of several ways of rendering the
graph). The prefix in the url used is "g" to avoid conflicting with the
"graph" folder. Really we should just move the graph folder outside
"app" and into "lib" but that would've meant touching a whole lot of
files which I didn't want to do
- Redirect default page to graph page
- Add `createRouteUrl` helper for constructing valid app urls
- Removed shift-click to open in sidebar (this was a feature for
prototyping and wasn't intended to be left in)
- Remove need for dynamic imports in page components