Fix initial data load (#297)
We were getting into bad initialization states because we wouldn't load
sync data from `/api/sync` when `/api/persist` returned `null`. This PR
fixes that in a hyper-minimalist way to quickly unbreak production.

We probably want to just remove the `/api/persist` stuff altogether but
that's a larger blast radius change and this only took 5 minutes.