More resilient approach to broadcast via Pusher (#305)
In Vercel logs I see two errors coming up from Pusher:

```
Unhandled Rejection: Error
  at new a (/var/task/.next/server/app/api/sync/route.js:5:9022)
  ...
  error: undefined,
  status: 413,
  body: 'The data content of this event exceeds the allowed maximum (10240 bytes). See https://pusher.com/docs/channels/server_api/http-api#publishing-events for more info\n'
```

and

```
Unhandled Rejection: Error
    at new a (/var/task/.next/server/app/api/sync/route.js:5:9022)
    ...
    type: 'system',
    errno: 'ECONNRESET',
    code: 'ECONNRESET'
```

These both come from an overly optimistic attempt to broadcast
transaction success.

The first comes from an oversized data body, which is addressed in this
PR by a new helper function that tries to split the `updates` array to
more manageable chunks when it sees things are oversized.

The second is just a standard connection error, and is addressed in this
PR with a simple retry loop that makes 3 attempts to push a message out
before giving up.

The two of these things together should make the sync code more
reliable. The first issue in particular was probably pretty insidious
for large graphs that we hadn't tested much.