fix: Await Pusher.trigger calls to ensure they get sent (#309)
Straightforward fix for a silly mistake: we need to `await` calls to
`Pusher.trigger` to ensure they get sent. This should fix the "sync 1
step behind" bug.

Note that reproducing this issue locally was always tricky. That makes
sense because of the different execution environments in Vercel
serverless vs running the API locally.