fix: Correctly update relation lists when reversing relation (#317)
Fix for tricky bug spotted by Phil + Taylor:

Whenever we switch from `Relation(id: r0, from: b, to: a)` to
`Relation(id: r0, from: a, to: b)`. the `a` node no longer has the `r0`
relation in it’s list.

The issue:
* in `setFrom`, we delete `r0` from the relation list of `b` and add it
to `a` (it’s actually already there, so this is a no-op)
* in `setTo`, we delete `r0` from the relation list of `a` and add it to
`b`. ← this is where we lose `r0` in the relation list for `a`

Fixed the logic here with an extra conditional + added test coverage for
it.