Add back ui for relation to relation (#310)
## Summary

Adds back support for relations to relations.

Completes ENT-3816

### Main changes

- Update `SearchAndReplaceDropdownPlugin.tsx` to properly support
selecting relations from the dropdown.
- Add `RelatedRelationView` to render a related relation. 
- The above includes a `TreeNodeInputSuffix` component, which
encapsulates some behaviour that's shared with nodes rendered as a
reference. In both of these cases, we don't have a lexical editor, which
is usually what allows the user to place selection inside a node and
handles actions like adding a sibling below or deleting. For components
that don't need a lexical editor, we place `TreeNodeInputSuffix` at the
end to allow focus and add support for relevant actions.

### Refactors

- We now have "RelatedNodeView" and "RelatedRelationView" components
which are responsible for rendering a node and relation respectively.

- The "RelatedNodeView" combines what was in the RelatedObjectEditor and
the "TreeNodeReference" that was previously in the RelatedObjectView.
Nothing important change in any of these components.
 
## How to test

- Create two nodes "1" and "2"
- Tab under "2" and type "test:" to label the relation
- Type "1" and you should get a dropdown (if you don't you need to turn
on "search and replace dropdown" in the settings). There should be an
option like "My graph -> child -> 1". Select it.
- You now have a relation pointing at another relation. The focus should
be at the end of the line.
- Pressing enter should add a sibling. Pressing backspace should place
the relation with a raw-text representation of it.