Add index for full-text search (#320)
## Summary

Add `CappedKeywordIndex`. This builds and maintains a trie of objects,
indexed by the first few characters of every word in the object text. I
capped the depth to a few characters cause I figured that amount of
filtering would be sufficient to get down to a small set which we can
then do a full text search on. On the 50k dataset, typing feels snappy
(sometimes a slight lag on the first character) compared to a 1-2s hang
before, at the cost of ~10% increase in memory.

Make it so SearchAndReplaceDropdownPlugin doesn't do the search for
matching nodes until after the user makes their first edit.

Part of ENT-3819

## How to test

On main:
- enable "search or replace" on all nodes
- import the 50k dataset:
https://drive.google.com/file/d/1FzKHRmFy_8Ic8utma34z_Op3fu8nH6AJ/view?usp=drive_link
- create a new node and start typing. There should be significant lag.

Then do the same on this branch. Typing should be snappy.

---------

Co-authored-by: RPC <67982490+rpc333@users.noreply.github.com>