Add back denormalized relation lists (#326)
I've mostly been able to figure the issue surfaced by the [Denormalized
relation lists PR](https://github.com/IdeaFlowCo/mew/pull/316). This PR
adds back denormalized relation lists and [a fix for the
issue](https://github.com/IdeaFlowCo/mew/pull/326/commits/0d751d0156178182395c89f328c42773f5b77119).
The only thing I haven't figured out is why this wasn't an issue before,
and what in the PR surfaced it. But I'm pretty confident that what I
found was a real existing issue, and that this PR fixes it.

The issue: `GraphStore.clear` would clear out our maps but leave the
references to the default objects untouched e.g.
`GraphStore.outlineRoot`, etc. Unless we make those properties optional,
we needed to point them at _something_. Later, when we load data from
the server, it would include serialized default objects, and because
those aren't in the graph store maps, we'd create new corresponding
objects and add those to the maps. Next we'd call
`ensureDefaultObjects`, which contains the code blocks to assign the
default objects to the corresponding store properties, but those code
blocks were skipped cause the objects were already in our maps.

So now we had separate instances of the default objects: in our maps and
on the store properties. The version assigned to the store property is
what drove the view, so the view would be empty. Pressing cmd+k would
add a child, but only to the version in our map, not the one driving the
view, so we wouldn't see anything on screen.

The fix was to call `ensureDefaultObjects` inside the `clear` method. By
doing so, we re-instantiate new versions of the default objects and add
them to both our maps and the store properties. Later when we load from
the server, the serialized default objects get applied to the existing
ones already in our maps, and since those instances are the same as the
ones referenced by our store properties, and since those drive the view,
now the view correctly reflects the store.