Command bar (#312)
### Command bar

This PR adds a command bar. Open it with `cmd+shift+k`. It supports
searching for and navigating to existing nodes or creating a new node.

When we jump to a node, we assign the top-most path of parents from the
node to the root as the path. If there is no path to the root, we just
jump to the node itself without a path.

From a product perspective, I'm not sure about this decision of using
the top-most path. Seems kinda arbitrary. Other options: 1) ignore the
path and just jump to the node without one or 2) only assign a path if
there's a single path of parents to the root, otherwise jump straight to
the node (cause in this case it's kinda "global"). It could be an issue
from a performance perspective too, cause we need to walk up the graph
of every result of the search (although if we limit the results maybe
it's fine).

To render the node path, it re-uses the code from the mention plugin,
which is now in a separate `Path.tsx` file.

Completes ENT-3820

### Refactor object paths

The PR also adds support for viewing nodes which don't have a path to
the outline root. Most of the code changes in the PR are actually for
that. Previously you'd specify the "root" of a view by providing a path
of relations from the outline/stream nodes to some other node. The node
which the path lands at is what you'd see as the header node of the
page.

This requirement that every node have a path from the root/stream is
unnecessary and undesired imo. It should be possible to create and view
a disconnected node. For example, when a new node is created through the
command bar, it's not related to anything. Even if we make a product
decision that nodes should always be connected to something, our app is
more resilient if we don't require that.

Now, we specify the "root" of a view by an object and a path of
relations to it is optional. The path doesn't need to start at the
outline or root. This is represented by `ObjectPath`. Our helper
functions for creating and parsing urls needed to be updated
accordingly. As did the interface for the Tree slightly.

### Misc
- Opted to remove state for sidebar outlines, rather than fix the type
issues, since we're not supporting that feature right now
- Add typechecking to ci tests