Agu/fix breadcrumb navigation and address mobile (#338)
## Summary

The latest update for a basic mobile version of Mew includes fixing the
breadcrumb navigation. It's now working well, and we've also resolved
some type errors related to navigation.

I added a hook to check for mobile devices in utils, but I'm not sure if
that's the right place for it.

Completes ENT-3836

## How to test

Test on mobile devices or screens narrower than 480px, and create some
indented nodes for navigation. It should display only the first and last
items, with a dropdown menu in the middle.

When trying to navigate to the page you're already on, it shouldn't
trigger any additional errors.