Fix parts of initial data load (#346)
The big thing with this PR: make sure that we're properly setting up
`GraphUpdates` for default objects (e.g. the outline root node).

Implementation notes:
- I re-established the `resetAndLoad` method because I think it's just
better to have slightly duplicated code paths that are easier to reason
about.
- I didn't write proper tests for the new method (yet) because the pain
here felt really urgent and I didn't want to block this PR on writing
some helper code to create good `SerializedGraphStore` objects to test
with, but that's a TODO.
- I'm queuing updates within `GraphStore.cleanup` because it doesn't
make sense to me to call `createDefaultObjects` without doing something
with the updates. I'm not sure if `cleanup` _should_ call
`createDefaultObjects`, but that's what's there currently and I don't
have a reason to change the behavior. Open to thoughts