Use getters for default nodes and relations (#349)
The default node + relation properties (`userRoot`, `outlineRoot`, etc)
created a foot gun where we had to make sure we were always keeping the
property values in sync with what was in `nodesById` and
`relationsById`. Swapping out the settable properties with getters
avoids all that pain.

I'm also cleaning up `createDefaultObjects` here:
* Make it private because I don't see a reasonable external use for it
* Rename it to `ensureDefaultObjectsCreated` because it doesn't always
create objects
* Make it queue updates directly in `updateManager` because that's what
we did every time we called it
* Stop returning the actual node/relation values because those were only
used to set the properties (which no longer exist)

I'm not sure if the new properties need to be `observable` or not. Also
pausing on merging this PR so that it can also incorporate any feedback
that would have gone into #346, which I merged early to unbreak some
things.