Agu/UI missing updates (#351)
## Summary

A bunch of minor fixes and tweaks:
- Command bar basic improvements. It's not properly "designed" yet, but
it has decent styling now.
- Tweaked the breadcrumb font size, which got a bit too large in UI 0.3
and wasn't adjusted in UI 0.4.
- The back arrow navigation on the right side of the sidebar was
confusing for me lately. I often mistook it for a button to hide the
sidebar. So, I moved it to the left to see if I could make its behavior
clearer.
- A few months ago, I implemented a simpler and prettier loading state,
but it got lost during refactoring. I've brought it back now

Completes ENT-3838, ENT-3899