Fix more overscroll issues and usability on mobile (#366)
- Finally fixed all the minor issues from my previous PR. Thank goodness
for "dvh" values - I totally forgot to use those for mobile.
- Additionally, I made the left node actions visible to improve
usability on mobile. This isn't ideal tho - the buttons are too small,
and we should refactor many things to create a proper mobile experience.
For now, though, the goal is simply to make it usable.

Testing involves creating about 30 nodes to see if the experience is
good so far. Scrolling up and down should be smooth, and everything
should be easily accessible.