Fix: Moving a bullet with cmd+shift+down freezes the app (#377)
The steps described in the ticket get you into an infinite loop of
focusing and blurring. I started trying to pin-point it down, but it was
a bit of a slog, so I just read through the code and did what made sense
instead, which seems to have fixed it.

The gist is that I added some if conditions to the handlers of
BindFocusToTreePlugin which check the current state of the tree
selection and editor before deciding if we should change the tree/editor
focus:

- On selection change, we check if there's mismatch b/w tree selection
and editor and if there is, we focus/blur the editor accordingly
- When the editor receives a blur/focus command, we check if the tree
selection and editor both say the current node is/isn't focused, and if
they agree, we update the tree selection according to the blur/focus
event.

Completes ENT-3920

(While reviewing #370 I noticed this PR fixes ENT-3951 too)

## How to test

See replication steps on ticket. On this branch, they don't cause
freezing anymore.
https://linear.app/ideaflow/issue/ENT-3920/moving-a-bullet-with-cmdshiftdown-freezes-the-app