Fix selection placement at start/end of editor (#378)
There's an issue right now where pressing right at the end of an editor
doesn't move the cursor to the start of the next line. See:
https://www.loom.com/share/5727d864030f4339a4424c43f9d68535

It looks to me it's the following code block introduced in #353 that's
the issue. I think it just sets the selection initially but won't
actually modify an existing one.

https://github.com/IdeaFlowCo/mew/blob/4e09d9c6aa35d3040acd3c718acea6c59e21bde4/src/app/editor/plugins/BindFocusToTreePlugin.tsx#L32-L38

This PR brings back the old implementation (plus an option which makes
the change synchronous) which fixes the issue shown in the video.

After doing that, extracting the editor position in the focus command handler 
is no longer needed as long as we allow undefined editor selection positions.
So that code block was removed.