Fix: Mentions are unable to select all relevant nodes (#380)
## Summary

Upon creating some node (say for example with text "Hello"), the next
node you create with text "He" should pick out the "Hello" node as one
of the autocompletions. However, that wasn't the case. I observed that
if there are N relevant nodes, then it would pick out only N-1 nodes.

The culprit seems to be an off-by-one error in a slice operation done on
the `limitedOptions` variable, which is addressed in the commit "No more
slicing".

Upon testing this doesn't seem to interfere with other functionality.
However, would still like to confirm with @eden-chan that this logic
doesn't interfere with the logic of this component.

## How to test

Create a node named "hello" and type in "he" in the next node. Before
this fix, the autocompletion for "he" would not show "hello" and only
the "create new node" option. Verify that the "he" node shows the
autocompletion for "hello".

## Checklist
(don't have access to the style guide)
- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).