fix: Bug where arrow navigation didn't work after tabbing away and back
We were setting the selection to null when tabbing away, but the editor
still had the selection. So when you tab back, you see the caret in the
editor, but from the apps perspective the selection is null, so the arrow
navigation does nothing.

The fix was just not to set the selection to null on blur. So the selection
is still set when you tab away and back. Maybe there are cases this leads
to unexpected behaviour, but I couldn't find any. Giving it a try.