fix: selection on click bug
There was an issue in the BindFocusToTreePlugin handler which updates
the tree selection to match the editor focus. It assumed that the focus
command handler would run *before* the editor became focused, and so wouldn't
set the selection unless the editor wasn't currently focused. But actually
this handler runs *after* the editor becomes focus, so we were never updating
the selection.

I removed the editor focus check and the selection updates as expected.

This condition was introduced in a PR fixing ENT-3920. In that PR, I wasn't
sure which change actually fixed the issue. Turns out it wasn't the condition
which we've removed in this commit, because after removing it, the ENT-3920
is *not* re-introduced.

Replication steps:
- create two bullets "a" and "b" below
- move selection up to "a" by pressing up arrow
- click on "b"
- pressing shift+up to initiate node-selection
- "a" becomes node-selected but it should be "b"