Notify user when the apps reloading cause there's a new version (#412)
See my explanation of the issue and suggestion for this change
[here](https://linear.app/ideaflow/issue/ENT-3862/app-reloads-mid-session#comment-08a27abd).

When a new build is deployed, nextjs automatically reloads the client on
the next page transition and loses app state.

Rather than fighting this nextjs behaviour, this PR just adds a message
that notifies the user that the app is reloading to a new version. The
mid-session loading will still be there, but at least the user will
understand why, and will be less likely to suspect it's a bug.

To do this, we inject the buildId into the client and then cache it in
local storage. On next start up, we compare the id in the cache to the
id in the current build, and notify the user in the loading message if
it's change.

It feels like there should be a better way to do this. Somehow nextjs is
detecting a diff in the build behind the scenes already, so it seems
weird that we need to inject the buildId and cache it to detect the same
thing. I couldn't find docs on this though, so maybe they just don't
expose that machinery. It could be that I missed it too though! It's
worth taking a look yourself.

Part of ENT-3862

## How to test

- Enable persistence and run the app `yarn build && yarn start`
- Open. You should see the "Loading..." screen. On refresh, you should
still see "Loading..."
- The build id is
[assigned](https://github.com/IdeaFlowCo/mew/blob/7df83aed70f0af688a632d533cce84e5fc5b592b/next.config.mjs#L14-L16)
to the latest commit hash. Add a new one with `git commit --allow-empty
-m "temp"`
- Stop the app server and start again with `yarn build && yarn start` 
- Now go back to the app and click on a bullet. The nextjs automatically
detects that a new build is being served (according to the new buildId)
and so refreshes itself (which causes loss of app state).
- You should see "A new version of Mew is loading..." on the loading
screen this time. That's what's been introduced by this PR.