fix: Do a full reset when fetching latest data from server (ENT-4023) (#422)
## Summary

We've been calling `load` as part of our snapshot refetch logic instead
of `resetAndLoad`, which prevents the client from "receiving" deletions
that it's missed, as the logic in `load` is strictly additive (only ever
creating nodes/relations/etc). The fix for that is straightforward: just
swap the function calls.

Addressing that did draw attention to some issues where the data in Tree
could potentially be stale, not reflecting the latest changes in
GraphStore. I fixed that by making `Tree.rootObject` be a computed value
that fetches the latest object from the GraphStore each time it's
needed. It seems to work fine, but I'm perpetually a little iffy when
modifying the Tree code, so would like any/all input on this change.

## How to test

Delete relation from database as Taylor does in video:
https://www.loom.com/share/ec1a5e8b47344f38adf315a48b3c286d