last fix for nail the dock icon size (#439)
fake it till you make it - I just wanted to get the right size for the
dock icon. Since it's not a native app, the icon doesn't match the real
size and we need to tweak the PNG.

To test this icon, you'll need to uninstall the current Mew PWA and
install it again.