Fix issue causing Jacob's account not to load (#442)
Jacob's account was completely failing to load. When you load his
account (see below), you can see that a lot of objects are placeholders.
There was at least one case where something was calling `isRoot` on
placeholder object, which would throw, because the placeholders `store`
property was set to null. In this PR, we pass in the store to the
placeholder object, which allows the app to load.

There's still an underlying issue here around placeholders and public
notes. I don't think we've really thought through how to handle cases
where you have lots of relations which refer to public object which
might not be available anymore (cause they were made private or
deleted). I'm leaving those issues for later though. Allow Jacob to load
the app is good enough for now.

Completes ENT-4073

## How to test

- Checkout
[taylor/fix-load](https://github.com/IdeaFlowCo/mew/tree/taylor/fix-load)
branch locally. On this branch, Jacob's account is hard-coded as the
currently logged in user (by passing auth).
- Update your .env.local to point to the verceldb (prod) 
- Run the app `yarn dev` and open
- The app will crash
- Now merge this branch `taylor/fix-jacob-load` onto `taylor/fix-load`
- Run and try again. It should load now.