fix splitting related relations (#445)
In a previous PR, we made it so the input suffix uses the same
`Tree.splitNode` method to handle enter events. This makes sure we have
consistent behaviours across components. But that method didn't handle
splitting related relations. This PR fixes that.

It also fixes an issue where the input suffix wasn't grabbing selection.
I pulled up the state which decides whether the editor or the suffix
should control selection, and pass that to both the editor and suffix.
Now they both condition on that, which fixed the issue.