Batch Syncing (fix for persistent upload included) [ENT-3879] (#419)
## Summary

This PR enables users to use the "Add to graph" upload functionality. 

Note: Currently, the "replace all data" function is broken due to an
independent problem (Error: Cannot delete relation from global to user),
which can be tackled after this PR. I believe it's due to a misuse of
the `resetAndReload` function in the importDialog component.

Features:
- chunked uploading with maximum row number
- basic loading dialog
- supports batches of types in the `const supportedTypes` variable

Some TODOs:
- updateRelationList: didn't implement it as I don't know how to format
for batch drizzle upsert - the orm syntax is weird with inserting a
batch of values with onConflictDoUpdate.
- createRelationTypes: there are always conflicts for this. relation
types are probably easy to duplicate. so I inserted with an
onConclictDoNothing.
- undoing local changes upon an error in uploading isn't done.
- didn't create new IDs for each conflict present in the DB.
- maximum filesize constraint

That being said though, this works well enough and I'm all out of time
for the trial week (10h/wk). Happy to finish this next week, just let me
know if I'm good to put more time towards it. This works for dbpedia-50k
and 1k.

Let me know what you all think of these changes :)

## How to test

Use the add to graph button in the importDialog component with dbpedia
1K or dbpedia 50k or both.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).