Only pull relation types when popover is open (#468)
Changes:
- Remove `get relationType` from graph store. Adding this getter causes
mobx to keep around an array of all the relation types, in addition to
the map we already have. Better to just do
`Object.values(graphStore.relationTypesById)` whenever we need the
array.
- In the relation combobox, don't pull the relation types until the
popover is open. Previously we'd do it whenever the combobox was
rendered.
- Use `graphStore.search` to pull the relation types which uses the trie
index.

Part of work to address performance problems after uploading
entanglement data.