Fire double colon handler whenever a colon is pressed next to a colon (#473)
We were already grabbing the text behind the caret position. Now we just
check that the text ends with a ":". If the user pressed ":" and a ":"
is behind it, then we assign the relation type.

The solves issue where user types ":", moves around (or sync causes
update), and then presses ":" next to it. Previous implementation
wouldn't fire, but Jacob expected it to.

Addresses
https://ideaflowteam.slack.com/archives/C07FU15QKTP/p1727306369123959