Refactor autocomplete and change hotkeys (#477)
Changes:
- Refactors the dropdown generally (completes ENT-4117). We no longer
duplicate and modify the lexical typeahead plugin code. We use a
separate component for the search-and-replace dropdowns and then can use
the lexical typeahead as-is. This gets rid of a lot of code and is
easier to reason about. [^1]
- When the passive autocomplete opens, don't have a default selection
set. This way when the user presses enter, it doesn't select an option.
If you *do* want to select an option, the user either arrows down to
select, or presses tab to select the first. See outline and discussion
of this behaviour
[here](https://ideaflowteam.slack.com/archives/C07GW72L9RU/p1727280283987349).
Completes ENT-4089

Also completes ENT-4106, ENT-4083, ENT-4046

## How to test

- Enabled "always" autocomplete. 
- Type in an editor. No option in the dropdown should be highlighted. 
- Press enter. The node should split.
- Now do the above again but press tab. This time the first option is
selected.
- When initiating the dropdown with ";" or "@", the first item _should_
be highlighted. It's only during passive autocomplete where no option is
highlighted.

[^1]: It's better but still not great tbh. There are some quirks of the
lexical typeahead plugin which we need to work around. If the
autocomplete gets any fancier, we probably should drop the
`LexicalTypeaheadMenuPlugin`. It's nice to keep right now cause it
handles a lot of tricky positioning logic for the @ dropdown. But we're
paying for that with harder to reason about state transition logic. If
the dropdown feature grows anymore, then that cost starts to outweigh
the benefit at which point we should drop `LexicalTypeaheadMenuPlugin`
and do the positioning ourselves.