Don't sync without auth (#482)
I tried doing an import while persistence was disabled and was getting
an error cause the update manager was trying to sync without
authedFetch. In this PR, I make it so the update manager is either
provided with an authenticated fetch or undefined. If it's undefined, it
skips all sync jobs.

Merging cause this is blocking some of Simon Zhang's work.