Gracefully handle mentions of deleted nodes (#494)
This was another thing causing crashes in Jacob's account if you ever
clicked into an editor that was mentioning a deleted node:
`SyncWithGraphPlugin` would go into an infinite update loop because
`graphNodeMatchesParagraph` could never return `true` when a mentioned
node was not available.

In general I think we just have to handle this case. Especially with our
increased emphasis on public graph editing, there's just too many cases
where a mention might get stale and we won't have had the chance to run
the reaction logic that prunes mentions to missing nodes.

I actually think the reaction logic is confusing/increases likelihood of
bugs, so I'm just deleting it here for now. We can add it back if we
really need it for some reason.

New rendering ends up looking like this

![image](https://github.com/user-attachments/assets/44aa7589-8294-4c66-a2d1-d98468dd54db)