When loading a circular relation, don't add it to the relation list twice (#497)
We were getting a lot of "Attempted to add item that is already in the
list" errors logged. Turns out the entanglement data includes some
circular relations. Maybe we shouldn't allow that? I'm not sure
how/where to enforce that though. The error would happen during loading.
For now, I'm just handling the case and avoiding adding the relation to
the relation list twice.