Avoid expensive renders (#499)
Turns out there are a few components that have a huge impact on render
time. The following reduces a ~400ms outline render to ~230ms. In all of
these cases, the components is fairly quick to render, but they're
included in every bullet, so they can really add up.

Changes:
- Don't render dropdown popover until you've hovered the ellipses (so
now only 1 is every rendered instead of N)
- Same thing for render combobox popover.
- Replace Tooltip with `title` attribute (which shows a plain tooltip).
It doesn't look as nice but it's way faster.