Avoid stale outline render (#498)
The way we were updating the outline root was causing the outline to
render with a stale state.

Here's what was happening:
- We'd use `router.push()` to set the url defining a new root
- This would cause `g/[[...page]]/page.tsx` to render. At this point,
the tree's root hasn't been updated, so the outline re-renders with the
current root.
- After that render finishes, the `g/[[...page]]/page.tsx` effect runs,
which sets the new tree's root, which causes another render.

The lag this would cause is especially noticeable when you're on a page
with a large number of bullets and you zoom in to a node with very few.
You'd expect this to be fast, but it's slow because the current page
re-renders before the new root is set and zoomed into.

In this PR, we only set the root based on the url during initial load.
After that, we stop setting the root reactively based on the path, which
avoids the outline root from rendering with a stale value. Instead we
use a `useSetRoot` helper which sets both the tree root (and renders
immediately) _and_ the url (to match the root we set, but doesn't cause
a render). In my testing, I often found render times were 3x faster when
zooming in/out from bullets.

---------

Co-authored-by: RPC <67982490+rpc333@users.noreply.github.com>